+++ test/zexbox/metrics_test.exs
@@ -2,9 +2,42 @@ defmodule Zexbox.MetricsTest do
   use ExUnit.Case
   alias Zexbox.Metrics
 
+  defmodule EchoEnricher do
+    @behaviour Zexbox.Metrics.ControllerSeriesEnricher
+    @impl true
+    def init(opts), do: {:initialised, opts}
+    @impl true
+    def call(series, _conn, _opts), do: series
+  end
+
+  setup do
+    on_exit(fn ->
+      :telemetry.detach("phoenix_controller_metrics")
+
+      try do
+        Supervisor.stop(Metrics)
+      catch
+        :exit, _reason -> :ok
+      end
+    end)
+
+    :ok
+  end
+
   test "start_link/1 starts the metrics supervisor" do
     {:ok, pid} = Metrics.start_link(nil)
     assert Process.alive?(pid)
+  end
+
+  test "start_link/1 normalises the enricher option through its init/1" do
+    {:ok, _pid} = Metrics.start_link(enricher: {EchoEnricher, [foo: :bar]})
+
+    handler =
+      [:phoenix, :endpoint, :stop]
+      |> :telemetry.list_handlers()
+      |> Enum.find(&(&1.id == "phoenix_controller_metrics"))
+
+    assert handler.config == %{enricher: {EchoEnricher, {:initialised, [foo: :bar]}}}
   end
 
   test "init/1 initializes the metrics supervisor" do